requirements.txt
aspose-words==24.5.0
